changed default health and attack type to double

+++ src/main/java/dungeonmania/movingEntity/MercBribedState.java
@@ -4,8 +4,8 @@ import dungeonmania.util.JSONConfig;
 
 public class MercBribedState implements MercenaryState {
 
-    private static final int DEFAULT_ALLY_HEALTH = JSONConfig.getConfig("ally_health");
-    private static final int DEFAULT_ALLY_ATTACK = JSONConfig.getConfig("ally_attack");
+    private static final double DEFAULT_ALLY_HEALTH = JSONConfig.getConfig("ally_health");
+    private static final double DEFAULT_ALLY_ATTACK = JSONConfig.getConfig("ally_attack");
 
     @Override
     public void currentState(Mercenary merc) {
