+++ src/main/java/dungeonmania/movingEntity/Spider.java
@@ -8,8 +8,8 @@ import dungeonmania.util.Position;
 
 public class Spider extends Enemy {
     
-    private static final int DEFAULT_SPIDER_HEALTH = JSONConfig.getConfig("spider_health");
-    private static final int DEFAULT_SPIDER_ATTACK = JSONConfig.getConfig("spider_attack");
+    private static final double DEFAULT_SPIDER_HEALTH = JSONConfig.getConfig("spider_health");
+    private static final double DEFAULT_SPIDER_ATTACK = JSONConfig.getConfig("spider_attack");
     private static final int DEFAULT_SPIDER_SPAWN_RATE = JSONConfig.getConfig("spider_spawn_rate");
 
     private boolean isClockwiseMove;
